move User Placemark on Map revision 1
